Review refactoring fix
